requirement.txt
Django==1.7.7
djangorestframework==2.3.14
psycopg2==2.5.3